django/requirements.txt
Django>=2.1.6
djangorestframework>=3.9.1
django-cors-headers==2.4.0
psycopg2
djangorestframework-jwt
cryptography
python-jose
pillow